+++ src/Joint_Grid.h
@@ -11,18 +11,18 @@
 #include <iostream>
 #include <algorithm>
 #include <assert.h>
-#include <climits>
 #include <math.h>
+#include <climits>
 
 #include "Clusters.h"
 
 using namespace std;
 
-vector<double> Find_1D_Grid(Cluster &clusters, int dim);
+vector< vector<double > > Find_Grid(Cluster &clusters, int min_bin_limit = 2);
 
-vector< vector<double > > Find_Grid(Cluster &clusters);
+vector<double> Find_1D_Grid(Cluster &clusters, int dim, int min_bin_limit);
 
-vector<vector<double> > prep_index(vector<double> &c1, vector<double> &c2, double median_1, double median_2);
+vector<vector<double> > prep_index(vector<double> &c1, vector<double> &c2, double middle_1, double middle_2);
 
 
 double binary_search_index(const vector<vector<double> > &c_index, const int left, const int right, const int size_c1,